api/requirements.txt
fastapi
uvicorn[standard]
pydantic
kafka-python
redis
qdrant-client>=1.9.0
google-generativeai
python-dotenv
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
bcrypt==4.0.1
torch
transformers
pandas
openai
python-dotenv
python-Levenshtein
numpy
elasticsearch<9.0.0
python-dateutil
apscheduler
requests